requirements.txt
python-resize-image
Pillow
numpy
scipy
argparse
scikit-learn==0.20.03
